backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "aerisun-backend"
version = "0.1.0"
description = "Aerisun v1 backend skeleton"
requires-python = ">=3.13"
dependencies = [
  "alembic>=1.18.4",
  "apscheduler>=3.10,<4",
  "bcrypt>=4.2.0",
  "fastapi>=0.135.1",
  "feedparser>=6.0",
  "httpx>=0.28",
  "pydantic>=2.12.5",
  "pydantic-settings>=2.8.0",
  "sentry-sdk[fastapi]>=2.0",
  "python-dateutil>=2.9",
  "python-multipart>=0.0.9",
  "pyyaml>=6.0",
  "slowapi>=0.1.9",
  "sqlalchemy>=2.0.48",
  "structlog>=24.0",
  "uvicorn[standard]>=0.34.0",
  "mcp>=1.12.4",
  "langgraph>=1.0.0",
  "langgraph-checkpoint-sqlite>=2.0.0",
  "aiosqlite>=0.20",
]

[dependency-groups]
dev = [
  "httpx>=0.28",
  "pytest>=8.3",
  "respx>=0.22",
  "ruff>=0.11",
]

[tool.setuptools]
package-dir = {"" = "src"}

[project.scripts]
aerisun-create-admin = "aerisun.cli:create_admin"
aerisun-reset-password = "aerisun.cli:reset_password"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-ra"
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["B008", "RUF001", "RUF002", "RUF003", "UP047"]

[tool.ruff.lint.isort]
known-first-party = ["aerisun"]

[tool.ruff.format]
quote-style = "double"
